front: transform computeBasePathSteps into computeBasePathStep
